feat: se reorganizó la estructura de importaciones para utilizar el nuevo módulo de base de datos
- Se creó el archivo database.py para manejar la inicialización de motores de base de datos.
- Se actualizaron las importaciones en varios módulos para referirse a la nueva ubicación de la base de datos.
