+++ app/shared/models/JIDs_model.py
@@ -10,7 +10,7 @@ from uuid import UUID
 from sqlalchemy import ForeignKey, Integer, Uuid
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from app.core.database import Base
+from app.core.db.database import Base
 
 if TYPE_CHECKING:
     from app.shared.models.auth_model import Auth
